sal cf

+++ src/engine/Strategy.java
@@ -6,7 +6,7 @@ import auxiliary.Pair;
 import auxiliary.Position;
 
 public interface Strategy {
-	public int eval(boolean player);
+	public double eval(boolean player);
 	public ArrayList<Pair<Position, Position>> nextMoves();
 	public int getDepth();
 }
